optimization: 优化 tasktmpl3 app 结构
