+++ gcloud/apigw/views/create_periodic_task.py
@@ -29,7 +29,7 @@ from gcloud.constants import PROJECT
 from gcloud.core.models import ProjectConfig
 from gcloud.periodictask.models import PeriodicTask
 from gcloud.taskflow3.models import TaskFlowInstance
-from gcloud.tasktmpl3.constants import NON_COMMON_TEMPLATE_TYPES
+from gcloud.constants import NON_COMMON_TEMPLATE_TYPES
 from gcloud.tasktmpl3.models import TaskTemplate
 from gcloud.apigw.views.utils import logger, info_data_from_period_task
 from gcloud.apigw.validators import CreatePriodicTaskValidator
