+++ gcloud/tasktmpl3/apis/tastypie/__init__.py
@@ -2,7 +2,7 @@
 """
 Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
 Edition) available.
-Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
+Copyright (C) 2017-2020 THL A29 Limited, a Tencent company. All rights reserved.
 Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 http://opensource.org/licenses/MIT
@@ -10,16 +10,3 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-
-PROJECT = "project"
-BUSINESS = "business"
-COMMON = "common"
-ONETIME = "onetime"
-
-TEMPLATE_EXPORTER_SOURCE_PROJECT = "project"
-TEMPLATE_EXPORTER_SOURCE_COMMON = "common"
-TEMPLATE_EXPORTER_VERSION = 1
-BIZ_INTERNAL_SET = "空闲机池"
-BIZ_INTERNAL_MODULE = ("空闲机", "待回收", "故障机")
-
-NON_COMMON_TEMPLATE_TYPES = {BUSINESS, PROJECT}
