+++ gcloud/tasktmpl3/apis/tastypie/resources.py
