+++ gcloud/tasktmpl3/domains/context.py
@@ -11,15 +11,24 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-PROJECT = "project"
-BUSINESS = "business"
-COMMON = "common"
-ONETIME = "onetime"
+import logging
 
-TEMPLATE_EXPORTER_SOURCE_PROJECT = "project"
-TEMPLATE_EXPORTER_SOURCE_COMMON = "common"
-TEMPLATE_EXPORTER_VERSION = 1
-BIZ_INTERNAL_SET = "空闲机池"
-BIZ_INTERNAL_MODULE = ("空闲机", "待回收", "故障机")
 
-NON_COMMON_TEMPLATE_TYPES = {BUSINESS, PROJECT}
+from gcloud.tasktmpl3.models import TaskTemplate
+
+
+logger = logging.getLogger("root")
+
+
+def get_template_context(pipeline_template, data_type, username=""):
+    try:
+        template = TaskTemplate.objects.get(pipeline_template=pipeline_template)
+    except TaskTemplate.DoesNotExist:
+        logger.warning("TaskTemplate Does not exist: pipeline_template.id=%s" % pipeline_template.pk)
+        return {}
+    context = {
+        "project_id": template.project.id,
+        "project_name": template.project.name,
+        "operator": template.pipeline_template.editor or username,
+    }
+    return context
