+++ gcloud/tests/tasktmpl3/domains/__init__.py
@@ -10,8 +10,3 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-
-from gcloud.constants import *  # noqa
-
-# history data compatible, some history data's template_source is still 'business'
-NON_COMMON_TEMPLATE_TYPES = {BUSINESS, PROJECT}
